An - Update models and handle empty modelLists

+++ Gateway/Services/Course/AdvisorHandlers/UpdateAdvisorHandler.cs
@@ -19,9 +19,9 @@ public class UpdateAdvisorHandler : RequestHandler<UpdateAdvisorCommand, Healpat
             if (entity is null)
                 return NotFound(string.Empty);
 
-            if (command.AddedMedias is not null)
+            if (command.AddedMedias is not null && command.AddedMedias.Count > 0)
                 _context.Multimedia.AddRange(command.AddedMedias);
-            if (command.RemovedMedias is not null)
+            if (command.RemovedMedias is not null && command.RemovedMedias.Count > 0)
                 await _context.Multimedia.DeleteExt(command.RemovedMedias);
             ApplyChanges(entity, command.Rq, command.UserId);
             await _context.SaveChangesAsync(cancellationToken);
