+++ Gateway/Services/Course/LectureHandlers/UpdateLectureHandler.cs
@@ -20,9 +20,9 @@ public class UpdateLectureHandler : RequestHandler<UpdateLectureCommand, Healpat
         ApplyChanges(entity, command);
         try
         {
-            if (command.AddedMedias is not null)
+            if (command.AddedMedias is not null && command.AddedMedias.Count > 0)
                 _context.Multimedia.AddRange(command.AddedMedias);
-            if (command.RemovedMedias is not null)
+            if (command.RemovedMedias is not null && command.RemovedMedias.Count > 0)
                 await _context.Multimedia.DeleteExt(command.RemovedMedias);
             await _context.SaveChangesAsync(cancellationToken);
             return Ok();
