+++ Gateway/Services/Library/ArticleHandlers/UpdateArticleHandler.cs
@@ -36,9 +36,9 @@ public class UpdateArticleHandler : RequestHandler<UpdateArticleCommand, Healpat
         {
             entity = ApplyChanges(entity, command, addedTags, removedTags);
 
-            if (command.AddedMedias is not null)
+            if (command.AddedMedias is not null && command.AddedMedias.Count > 0)
                 _context.Multimedia.AddRange(command.AddedMedias);
-            if (command.RemovedMedias is not null)
+            if (command.RemovedMedias is not null && command.RemovedMedias.Count > 0)
                 await _context.Multimedia.DeleteExt(command.RemovedMedias);
 
             await _context.SaveChangesAsync(cancellationToken);
